pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepphase-x"
version = "0.1.0"
description = "XFEL Weak Signal Enhancement & Phase Retrieval System"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "DeepPhase-X Team"}
]
keywords = ["XFEL", "CDI", "deep learning", "phase retrieval", "denoising"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "torch>=2.0.0",
    "h5py>=3.7.0",
    "biopython>=1.79",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "hypothesis>=6.0.0",
    "black",
    "isort",
    "mypy",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88
